cambio del campo cliente por biginteger

+++ lugrascolv2/facturacion/models.py
@@ -33,7 +33,7 @@ class Facturas(models.Model):
     fecha_facturacion = models.DateField()
     total_factura = models.FloatField()
     id_orden_field = models.IntegerField(db_column='id_orden ')    # Field renamed to remove unsuitable characters. Field renamed because it ended with '_'.
-    cliente = models.IntegerField()
+    cliente = models.BigIntegerField()
     estado = models.TextField()
     ica= models.BooleanField(default=False)
